Add files via upload
Mostly have finished up on the Login Page (intuitively labeled the documents). Registration Page and the Seed Round. (Although it still lacks a User Interface). Note that the website will not look pretty as this is mainly PHP. For program to work setup a MYSQL Database. Ideally using WAMP which is what I do. Exporting seems to work although as of right now I can only export the Entire Database of debaters of a particular tournament and not the wins or the speaks but this should be easy and would be a simple add-on that I am looking forward to integrate into the code.