Added Margins into the system + Progression
Added Margins into the system after I realized that it was missing from the end system. Also added the Button that makes you able to progress from the rounds into another after you have entered the details of the wins of each round. Validation() function still has some bugs that needs to be dealt with.