The Final Commit
This concludes the final commit for this software; it was amazing while it lasted. I ran this software at a tournament and other than the human errors I noticed one massive mistake; I had done the fold all wrong. I corrected the fold in this program, was some minor algorithm adjustments but it is now done. While this software may not be used widely by all the tournaments in the world this source code will always exists on the Web as an OpenSource Algorithm to be included into anyone who dares ventures these depths. Thanks for the journey!